requirements.txt
matplotlib==3.10.1
numpy==1.26.4
opencv-python==4.11.0.86
pandas==2.2.3
tqdm==4.67.1
faiss-cpu
torch
torchvision
opencv-python
streamlit